Add verification of existance befor edition, or suppression

+++ src/controller/BackController.php
@@ -95,58 +95,66 @@ class BackController extends Controller
     public function editArticle(Parameter $post, $articleId)
     {
         $this->checkAdmin();
-        $article = $this->articleDAO->getArticle($articleId);
-        PictureManager::findArticlePictures($article);
-        if ($post->get('publish') || $post->get('draft') || $post->get('toCorrect')) {
-            $errors = $this->validation->validate($post, 'Article');
+        if ($this->articleDAO->checkArticle($articleId)) {
+            $article = $this->articleDAO->getArticle($articleId);
+            PictureManager::findArticlePictures($article);
+            if ($post->get('publish') || $post->get('draft') || $post->get('toCorrect')) {
+                $errors = $this->validation->validate($post, 'Article');
 
-            if ($article->getTitle() !== $post->get('title')) {
-                if ($this->articleDAO->checkArticleTitle($post)) {
-                    $errors['unique'] = $this->articleDAO->checkArticleTitle($post);
+                if ($article->getTitle() !== $post->get('title')) {
+                    if ($this->articleDAO->checkArticleTitle($post)) {
+                        $errors['unique'] = $this->articleDAO->checkArticleTitle($post);
+                    }
                 }
-            }
-            if (!$errors) {
-                if ($post->get('publish')) {
-                    $status = 1;
-                }
-                if ($post->get('toCorrect')) {
-                    $status = 2;
-                }
-                if ($post->get('draft')) {
-                    $status = 0;
-                }
+                if (!$errors) {
+                    if ($post->get('publish')) {
+                        $status = 1;
+                    }
+                    if ($post->get('toCorrect')) {
+                        $status = 2;
+                    }
+                    if ($post->get('draft')) {
+                        $status = 0;
+                    }
 
-                $this->articleDAO->editArticle($post, $articleId, $this->session->get('id'), $status);
-                $this->session->set('success_message', '<strong>L\' article a bien été modifié</strong>');
-                header('Location: index.php?route=administration');
-                exit();
+                    $this->articleDAO->editArticle($post, $articleId, $this->session->get('id'), $status);
+                    $this->session->set('success_message', '<strong>L\' article a bien été modifié</strong>');
+                    header('Location: index.php?route=administration');
+                    exit();
+                } else {
+                    $this->session->set('error_message', '<strong>L\' article n\'a pas été modifié</strong> ' . $errors['unique'] . $errors['title'] . $errors['content']);
+                    return $this->view->render('edit_article', [
+                        'post' => $post,
+                        'errors' => $errors
+                    ]);
+                }
             } else {
-                $this->session->set('error_message', '<strong>L\' article n\'a pas été modifié</strong> ' . $errors['unique'] . $errors['title'] . $errors['content']);
+                $post->set('id', $article->getId());
+                $post->set('title', $article->getTitle());
+                $post->set('picture_file_name', $article->getPictureFileName());
+                $post->set('picture', $article->getPicture());
+                $post->set('thumbail', $article->getThumbail());
+                $post->set('content', $article->getContent());
+                $post->set('author', $article->getAuthor());
+
                 return $this->view->render('edit_article', [
                     'post' => $post,
-                    'errors' => $errors
                 ]);
             }
-        } else {
-            $post->set('id', $article->getId());
-            $post->set('title', $article->getTitle());
-            $post->set('picture_file_name', $article->getPictureFileName());
-            $post->set('picture', $article->getPicture());
-            $post->set('thumbail', $article->getThumbail());
-            $post->set('content', $article->getContent());
-            $post->set('author', $article->getAuthor());
-
-            return $this->view->render('edit_article', [
-                'post' => $post,
-            ]);
         }
+        header('Location: index.php?route=administration');
+        exit();
     }
 
     public function deleteArticle($articleId)
     {
         $this->checkAdmin();
-        $this->articleDAO->deleteArticle($articleId);
-        $this->session->set('success_message', '<strong>L\'article a bien été supprimé</strong>');
+        if ($this->articleDAO->checkArticle($articleId)) {
+            $this->articleDAO->deleteArticle($articleId);
+            $this->session->set('success_message', '<strong>L\'article a bien été supprimé</strong>');
+        } else {
+            $this->session->set('error_message', '<strong>Supression impossible</strong>');
+        }
         header('Location: index.php?route=administration');
         exit();
     }
@@ -157,7 +165,7 @@ class BackController extends Controller
         $userId = $this->session->get('id');
         $user = $this->userDAO->getUser($userId);
         PictureManager::findAvatar($user);
-        return $this->view->render('profile',[
+        return $this->view->render('profile', [
             'user' => $user
         ]);
     }
@@ -200,36 +208,56 @@ class BackController extends Controller
     public function approveComment($commentId)
     {
         $this->checkAdmin();
-        $this->commentDAO->approveComment($commentId);
-        $this->session->set('success_message', '<strong>Commentaire approuvé</strong>');
-        header("Location: " . $_SERVER["HTTP_REFERER"]);
+        if ($this->commentDAO->checkComment($commentId)) {
+            $this->commentDAO->approveComment($commentId);
+            $this->session->set('success_message', '<strong>Commentaire approuvé</strong>');
+            header("Location: " . $_SERVER["HTTP_REFERER"]);
+            exit();
+        }
+        $this->session->set('error_message', '<strong>Approbation impossible</strong>');
+        header('Location: index.php?route=roman');
         exit();
     }
 
     public function deleteComment($commentId)
     {
         $this->checkAdmin();
-        $this->commentDAO->deleteComment($commentId);
-        $this->session->set('success_message', '<strong>Commentaire supprimé</strong>');
-        header("Location: " . $_SERVER["HTTP_REFERER"]);
+        if ($this->commentDAO->checkComment($commentId)) {
+            $this->commentDAO->deleteComment($commentId);
+            $this->session->set('success_message', '<strong>Commentaire supprimé</strong>');
+            header("Location: " . $_SERVER["HTTP_REFERER"]);
+            exit();
+        }
+        $this->session->set('error_message', '<strong>Suppression impossible</strong>');
+        header('Location: index.php?route=roman');
         exit();
     }
 
     public function archiveComment($commentId)
     {
         $this->checkAdmin();
-        $this->commentDAO->archiveComment($commentId);
-        $this->session->set('success_message', '<strong>Commentaire archivé</strong>');
-        header("Location: " . $_SERVER["HTTP_REFERER"]);
+        if ($this->commentDAO->checkComment($commentId)) {
+            $this->commentDAO->archiveComment($commentId);
+            $this->session->set('success_message', '<strong>Commentaire archivé</strong>');
+            header("Location: " . $_SERVER["HTTP_REFERER"]);
+            exit();
+        }
+        $this->session->set('error_message', '<strong>Archivage impossible</strong>');
+        header('Location: index.php?route=roman');
         exit();
     }
 
     public function hideComment($commentId)
     {
         $this->checkAdmin();
-        $this->commentDAO->hideComment($commentId);
-        $this->session->set('success_message', '<strong>Commentaire masqué</strong>');
-        header("Location: " . $_SERVER["HTTP_REFERER"]);
+        if ($this->commentDAO->checkComment($commentId)) {
+            $this->commentDAO->hideComment($commentId);
+            $this->session->set('success_message', '<strong>Commentaire masqué</strong>');
+            header("Location: " . $_SERVER["HTTP_REFERER"]);
+            exit();
+        }
+        $this->session->set('error_message', '<strong>Masquage impossible</strong>');
+        header('Location: index.php?route=roman');
         exit();
     }
 }
